+++ package.json
@@ -11,14 +11,14 @@
     "format": "prettier --write ."
   },
   "dependencies": {
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
     "@fontsource/poppins": "^5.0.13",
     "@hookform/resolvers": "^3.3.4",
-    "@mui/icons-material": "^5.15.15",
-    "@mui/material": "^5.15.15",
+    "@mui/icons-material": "^5.18.0",
+    "@mui/material": "^5.18.0",
     "@mui/styled-engine-sc": "^6.0.0-alpha.18",
-    "@mui/system": "^5.15.14",
+    "@mui/system": "^5.18.0",
     "@reduxjs/toolkit": "^2.2.2",
     "camelcase-keys": "^9.1.3",
     "jwt-decode": "^4.0.0",
@@ -31,8 +31,8 @@
     "yup": "^1.4.0"
   },
   "devDependencies": {
-    "@types/react": "^18.2.66",
-    "@types/react-dom": "^18.2.22",
+    "@types/react": "^18.3.23",
+    "@types/react-dom": "^18.3.7",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
     "@vitejs/plugin-react": "^4.2.1",
@@ -42,7 +42,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.6",
     "prettier": "^3.5.3",
-    "typescript": "^5.2.2",
+    "typescript": "^5.8.3",
     "vite": "^6.2.0"
   }
 }
