+++ src/LandingPage/Navigation.tsx
@@ -1,23 +1,15 @@
 import AppBar from "@mui/material/AppBar";
 import Toolbar from "@mui/material/Toolbar";
 import Button from "@mui/material/Button";
-import MenuIcon from "@mui/icons-material/Menu";
-import IconButton from "@mui/material/IconButton";
 import { useNavigate } from "react-router-dom";
 import { ROUTE } from "@/config/route.config.ts";
 import { styled } from "@mui/material/styles";
 import Box from "@mui/material/Box";
 import Logo from "@/components/Logo/Logo";
-import { useState } from "react";
-import NaviPanel from "@/components/NaviPanel/NaviPanel";
+import Hamburger from "@/components/Buttons/Hamburger";
 
 const Navigation = () => {
-  const [drawerOpen, setDrawerOpen] = useState(false);
   const navigate = useNavigate();
-
-  const toggleDrawer = (open: boolean) => () => {
-    setDrawerOpen(open);
-  };
 
   const handleLoginClick = () => {
     navigate(ROUTE.LOGIN);
@@ -27,15 +19,7 @@ const Navigation = () => {
     <NavigationContainer>
       <StyledAppBar>
         <Toolbar>
-          <StyledMenuIcon
-            size="large"
-            edge="start"
-            color="inherit"
-            aria-label="menu"
-            onClick={toggleDrawer(true)}
-          >
-            <MenuIcon />
-          </StyledMenuIcon>
+          <Hamburger />
           <Logo />
           <LogInButton
             variant="contained"
@@ -47,7 +31,6 @@ const Navigation = () => {
           </LogInButton>
         </Toolbar>
       </StyledAppBar>
-      <NaviPanel open={drawerOpen} onClose={toggleDrawer(false)} />
     </NavigationContainer>
   );
 };
@@ -59,10 +42,6 @@ const NavigationContainer = styled(Box)({
 const StyledAppBar = styled(AppBar)(({ theme }) => ({
   backgroundColor: theme.palette.secondary.main,
 }));
-
-const StyledMenuIcon = styled(IconButton)({
-  marginRight: 2,
-});
 
 const LogInButton = styled(Button)(({ theme }) => ({
   backgroundColor: theme.palette.button.main,
